Update to php7.2, install ojs as part of image
So far, the installation actually used to happen on the container
startup. This meant building the image was not doing enough.
The installation of ojs has been moved to the Dockerfile itself so
that the images once built can be easily deployed.

Minor changes to git commands were made to make things far
faster.